Gutted the b1pi_one.csh script and replaced it with a source of the setenv.csh script in the sim-recon build being used.
There was a mismatch in the environments setup done by the
build_halld.csh and b1pi_one.csh scripts causing b1pi_one.csh
to fail. Specifically, because the correct XERCES library was
not in the path.

The setenv.csh script (rather, the platform-specific one) is
generated by the build_halld.csh to reflect the important
environment variables used when it was run. Sourceing this
in b1pi_one.csh guarantees consistency between the two.